Bug 2020608 - Part 15: Require unsigned integral types for CeilingLog2 and FindMostSignificantBit. r=sergesanspaille,nbp
Instead of implicitly casting the input to unsigned, require this happens explicitly
in the caller. This aligns with `<bit>` std-functions and `Rotate{Left,Right}` which
also require unsigned inputs.

Differential Revision: https://phabricator.services.mozilla.com/D286324

+++ js/src/jit/x86/CodeGenerator-x86.cpp
@@ -836,7 +836,7 @@ void CodeGenerator::visitMulI64(LMulI64* lir) {
       default:
         if (constant > 0) {
           // Use shift if constant is power of 2.
-          int32_t shift = mozilla::FloorLog2(constant);
+          int32_t shift = mozilla::FloorLog2(uint64_t(constant));
           if (int64_t(1) << shift == constant) {
             masm.lshift64(Imm32(shift), lhs);
             return;
@@ -1324,7 +1324,7 @@ void CodeGenerator::visitMulIntPtr(LMulIntPtr* ins) {
 
     // Use shift if constant is a power of 2.
     if (constant > 0 && std::has_single_bit(uintptr_t(constant))) {
-      uint32_t shift = mozilla::FloorLog2(constant);
+      uint32_t shift = mozilla::FloorLog2(uintptr_t(constant));
       masm.lshiftPtr(Imm32(shift), lhs);
       return;
     }
